+++ src/utils/float_cmp.c
@@ -6,13 +6,13 @@
 /*   By: cado-car <cado-car@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/04 19:44:03 by cado-car          #+#    #+#             */
-/*   Updated: 2023/04/08 15:03:15 by cado-car         ###   ########.fr       */
+/*   Updated: 2023/04/10 15:04:37 by cado-car         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-bool	float_cmp(float a, float b, float epsilon)
+bool	float_cmp(double a, double b, double epsilon)
 {
 	if (abs_float(a - b) < epsilon)
 		return (true);
